Board now has relationship with blocks + Layer completion
There is now an array representation of the board spaces, which holds
the 'block' objects (instead of a list). This way you can quickly access
a block at a given position on the board. The blocks fall in a
contiguous manner to cover each space on the board. When a layer fills
up it is destroyed. (At the moment only the bottom layer of bricks can
be destroyed).